refactor: 코드 리뷰 반영(일일 충전 한도 수정)
- DB에 별도의 테이블로 관리할까 고민했지만 Redis에서 관리합니다.
- 별도의 테이블을 생성하면 디스크에 저장 공간이 필요하고 클러스터 인덱스를 위한 메모리 공간이 필요합니다.
- 또한 이를 위한 DB connection도 필요하기에 DB 부담은 최소화 하려고 했습니다.
- 그래서 DB에서 한 번 읽은 후에는 Redis에 <PK, 일일한도>와 같이 저장하도록 했습니다.
- 일정 시간이 지나면 안쓰게 되지만 DB를 사용했을 때 인덱스 테이블이 사용하는 메모리 크기와 큰 차이가 없다고 생각합니다.

+++ src/main/java/org/c4marathon/assignment/bankaccount/dto/response/MainAccountResponseDto.java
@@ -4,7 +4,7 @@ import org.c4marathon.assignment.bankaccount.entity.MainAccount;
 
 public record MainAccountResponseDto(
 	long accountPk,
-	int chargeLimit,
+	long chargeLimit,
 	long money
 ) {
 	public MainAccountResponseDto(MainAccount mainAccount) {
